requirements/base.txt
alabaster==0.7.12
amqp==2.5.2
appdirs==1.4.3
argon2-cffi==19.2.0
asgiref==3.2.7
astroid==2.3.3
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
black==19.10b0
bs4==0.0.1
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
click==7.1.1
coverage==5.1
dateparser==0.7.4
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.5
django-allauth==0.41.0
django-celery-beat==2.0.0
django-coverage-plugin==1.8.0
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.9
django-model-utils==4.0.0
django-redis==4.11.0
django-stubs==1.5.0
django-timezone-field==4.0
djangorestframework==3.11.0
docutils==0.16
entrypoints==0.3
factory-boy==2.12.0
Faker==4.0.3
filelock==3.0.12
flake8==3.7.9
flake8-isort==3.0.0
gunicorn==20.0.4
identify==1.4.14
idna==2.9
imagesize==1.2.0
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
kombu==4.6.8
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.770
mypy-extensions==0.4.3
nodeenv==1.3.5
oauthlib==3.1.0
packaging==20.3
pandas==1.0.4
parso==0.7.0
pathspec==0.8.0
pdfminer.six==20200104
pdfplumber==0.5.21
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
pluggy==0.13.1
pre-commit==2.2.0
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.8.1
pycodestyle==2.5.0
pycparser==2.20
pycryptodome==3.9.7
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.4.4
pylint-celery==0.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==5.3.5
pytest-django==3.9.0
pytest-sugar==0.9.2
python-crontab==2.4.1
python-dateutil==2.8.1
python-slugify==4.0.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
soupsieve==2.0.1
Sphinx==3.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
termcolor==1.1.0
testfixtures==6.14.1
text-unidecode==1.3
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
tzlocal==2.1
unicodecsv==0.14.1
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.18
Wand==0.6.1
wcwidth==0.1.9
Werkzeug==1.0.1
whitenoise==5.0.1
wrapt==1.11.2
dateparser==0.7.4
